Rename existing Select classes to StaticSelect
