+++ src/BlockElement.php
@@ -23,7 +23,7 @@ abstract class BlockElement extends Payload
         'datepicker'                 => ['DatePicker',               ['section', 'actions', 'input']],
         'timepicker'                 => ['Timepicker',               ['section', 'actions', 'input']],
         'image'                      => ['Image',                    ['section', 'context']],
-        'multi_static_select'        => ['MultiSelect',              ['section', 'input']],
+        'multi_static_select'        => ['MultiStaticSelect',        ['section', 'input']],
         'multi_external_select'      => ['MultiExternalSelect',      ['section', 'input']],
         'multi_users_select'         => ['MultiUsersSelect',         ['section', 'input']],
         'multi_conversations_select' => ['MultiConversationsSelect', ['section', 'input']],
@@ -31,7 +31,7 @@ abstract class BlockElement extends Payload
         'overflow'                   => ['Overflow',                 ['section', 'actions']],
         'plain_text_input'           => ['TextInput',                ['input']],
         'radio_buttons'              => ['RadioButtons',             ['section', 'actions', 'input']],
-        'static_select'              => ['Select',                   ['section', 'actions', 'input']],
+        'static_select'              => ['StaticSelect',             ['section', 'actions', 'input']],
         'external_select'            => ['ExternalSelect',           ['section', 'actions', 'input']],
         'users_select'               => ['UsersSelect',              ['section', 'actions', 'input']],
         'conversations_select'       => ['ConversationsSelect',      ['section', 'actions', 'input']],
