+++ src/BlockElement/AbstractSelect.php
@@ -1,0 +1,9 @@
+<?php
+namespace Maknz\Slack\BlockElement;
+
+/**
+ * @deprecated This class is deprecated in favour of AbstractStaticSelect since 2.2.0 and will be removed in a future release.
+ */
+abstract class AbstractSelect extends AbstractStaticSelect
+{
+}
