+++ tests/Block/ContextUnitTest.php
@@ -38,11 +38,11 @@ class ContextUnitTest extends TestCase
     public function testInvalidElement()
     {
         $this->expectException(InvalidArgumentException::class);
-        $this->expectExceptionMessage('Block element Maknz\Slack\BlockElement\MultiSelect is not valid');
+        $this->expectExceptionMessage('Block element Maknz\Slack\BlockElement\MultiStaticSelect is not valid');
         $c = new Context([
             'elements' => [[
                 'type' => 'multi_static_select',
-                'placeholder' => 'MultiSelect placeholder',
+                'placeholder' => 'MultiStaticSelect placeholder',
                 'options' => [[
                     'text'  => 'Option 1',
                     'value' => 'option_1',
