+++ tests/BlockElementUnitTest.php
@@ -26,7 +26,7 @@ class BlockElementUnitTest extends TestCase
             ['datepicker',                 BlockElement\DatePicker::class],
             ['timepicker',                 BlockElement\Timepicker::class],
             ['image',                      BlockElement\Image::class],
-            ['multi_static_select',        BlockElement\MultiSelect::class],
+            ['multi_static_select',        BlockElement\MultiStaticSelect::class],
             ['multi_external_select',      BlockElement\MultiExternalSelect::class],
             ['multi_users_select',         BlockElement\MultiUsersSelect::class],
             ['multi_conversations_select', BlockElement\MultiConversationsSelect::class],
@@ -34,7 +34,7 @@ class BlockElementUnitTest extends TestCase
             ['overflow',                   BlockElement\Overflow::class],
             ['plain_text_input',           BlockElement\TextInput::class],
             ['radio_buttons',              BlockElement\RadioButtons::class],
-            ['static_select',              BlockElement\Select::class],
+            ['static_select',              BlockElement\StaticSelect::class],
             ['external_select',            BlockElement\ExternalSelect::class],
             ['users_select',               BlockElement\UsersSelect::class],
             ['conversations_select',       BlockElement\ConversationsSelect::class],
